Feat: create Greeting component
