+++ src/App.js
@@ -1,21 +1,15 @@
-import { useSelector, useDispatch } from 'react-redux';
-import { useEffect } from 'react';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
 import './App.css';
 
-import { fetchGreeting } from './redux/greetings/reducers/greetingsReducer';
+import Greeting from './components/greeting';
 
 function App() {
-  const dispatch = useDispatch();
-  const greeting = useSelector(({ greetingReducer }) => greetingReducer.messages);
-
-  useEffect(() => {
-    dispatch(fetchGreeting());
-  }, []);
-
   return (
-    <div className="App">
-      <h1>{greeting.length > 0 && greeting[0].result.message}</h1>
-    </div>
+    <Router>
+      <Routes>
+        <Route path="/" element={<Greeting />} />
+      </Routes>
+    </Router>
   );
 }
